make the code compile and run on EM64T(~AMD64)
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@204 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cv/cvconvhull.cpp
@@ -210,7 +210,7 @@ icvCalcAndWritePtIndices( CvPoint** pointer, int* stack, int start, int end,
     {
         CvPoint* ptr = (CvPoint*)pointer[stack[i]];
         CvSeqBlock* block = ptseq->first;
-        while( (unsigned)(idx = ptr - (CvPoint*)block->data) >= (unsigned)block->count )
+        while( (unsigned)(idx = (int)(ptr - (CvPoint*)block->data)) >= (unsigned)block->count )
         {
             block = block->next;
             if( block == ptseq->first )
